requirements.txt
streamlit
qrcode
firebase_admin
streamlit_geolocation
streamlit_folium
leafmap
geopy
taxicab
osmnx
networkx
google.generativeai